electronics-service/requirements.txt
django
djangorestframework
requests
psycopg2-binary
django-filter
pika
